// abp-vnext-pro-basic-no-ocelot-nuget/vben28/src/components/Table/src/types/table.ts
export declare type SortOrder = 'ascend' | 'descend';
// ...
export interface SorterResult {
  column: ColumnProps;
  order: SortOrder;
  field: string;
  columnKey: string;
}

export interface FetchParams {
  searchInfo?: Recordable;
  page?: number;
  sortInfo?: Recordable;
  filterInfo?: Recordable;
}